+++ app/src/main/java/com/pavlovalexey/torpedo/ui/MenuFragment.kt
@@ -7,15 +7,17 @@ import android.view.ViewGroup
 import android.widget.Button
 import android.widget.Toast
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
 import com.pavlovalexey.torpedo.R
 import com.pavlovalexey.torpedo.viewmodel.MenuViewModel
 
 /** фрагмент для верхнего меню игры. Содержит справку по ресурсам и события, кнопки*/
 
+
+import androidx.fragment.app.activityViewModels
+
 class MenuFragment : Fragment() {
 
-    private val viewModel: MenuViewModel by viewModels()
+    private val viewModel: MenuViewModel by activityViewModels()
 
     private fun toast(message: String) {
         Toast.makeText(requireContext(), message, Toast.LENGTH_SHORT).show()
@@ -32,31 +34,30 @@ class MenuFragment : Fragment() {
         val settingsButton = view.findViewById<Button>(R.id.settingsButton)
         val closeSaveButton = view.findViewById<Button>(R.id.closeSaveButton)
         val menuLayout = view.findViewById<View>(R.id.fragment_menu)
-        (activity as MainActivity).hideOptionsLayout()
+//        (requireActivity() as MainActivity).hideOptionsLayout()
 
         soundButton.setOnClickListener {
             toast("Вы нажали кнопку 'Звук' Функция в разработке")
-            toast("Функция в разработке")
         }
 
         autoPlayButton.setOnClickListener {
-            toast("Вы нажали кнопку 'Автоплей'")
-            toast("Функция в разработке")
+            toast("Вы нажали кнопку 'Автоплей' Функция в разработке")
         }
 
         settingsButton.setOnClickListener {
-            toast("Вы нажали кнопку 'Настройки'")
-            toast("Функция в разработке")
+            toast("Вы нажали кнопку 'Настройки' Функция в разработке")
         }
 
         closeSaveButton.setOnClickListener {
-            toast("Вы нажали кнопку 'Закрыть + Сохранить Прогресс'")
-            toast("Функция в разработке")
+            toast("Вы нажали кнопку 'Закрыть + Сохранить Прогресс' Функция в разработке")
         }
 
         menuLayout.setOnClickListener {
-            (activity as MainActivity).showOptionsLayout()
-            parentFragmentManager.beginTransaction().remove(this@MenuFragment).commit()
+//            (requireActivity() as MainActivity).showOptionsLayout()
+//            parentFragmentManager.beginTransaction().remove(this@MenuFragment).commit()
+
+            val fragmentManager = parentFragmentManager
+            fragmentManager.popBackStack()
         }
 
         return view
