+++ app/src/main/res/layout/activity_main.xml
@@ -1,135 +1,21 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/rootLayout"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    tools:context=".ui.MainActivity">
 
-    <FrameLayout
-        android:id="@+id/sceneContainer"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
+    <fragment
+        android:id="@+id/nav_host_fragment"
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/nav_graph"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
-        <ImageView
-            android:id="@+id/sceneImageView"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@drawable/newtag"
-            android:scaleType="centerCrop" />
-    </FrameLayout>
-
-    <!-- Верхняя панель с ресурсами -->
-    <LinearLayout
-        android:id="@+id/resourcesLayout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@drawable/horizontal_gradient"
-        android:orientation="horizontal"
-        android:padding="0dp">
-
-        <TextView
-            android:id="@+id/rublesTextView"
-            style="@style/Resurses"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="" />
-
-        <TextView
-            android:id="@+id/fameTextView"
-            style="@style/Resurses"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="" />
-
-        <TextView
-            android:id="@+id/capitalTextView"
-            style="@style/Resurses"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="" />
-
-        <TextView
-            android:id="@+id/necronomiconTextView"
-            style="@style/Resurses"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="" />
-    </LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/resourcesLayout2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/resourcesLayout"
-        android:background="@drawable/horizontal_gradient_light"
-        android:orientation="horizontal"
-        android:padding="0dp">
-
-        <TextView
-            android:id="@+id/teamLoyaltyTextView"
-            style="@style/Resurses"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="" />
-
-        <TextView
-            android:id="@+id/vodkaTextView"
-            style="@style/Resurses"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="" />
-
-        <TextView
-            android:id="@+id/maximTextView"
-            style="@style/Resurses"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="" />
-
-
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/relationshipTextView"
-        style="@style/Resurses"
-        android:layout_width="@dimen/dp44"
-        android:layout_height="@dimen/dp44"
-        android:layout_below="@+id/resourcesLayout2"
-        android:layout_alignParentStart="true"
-        android:gravity="center"
-        android:text="" />
-
-    <TextView
-        android:id="@+id/dialogueTextView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_above="@id/optionsLayout"
-        android:layout_centerHorizontal="true"
-        android:background="@drawable/horizontal_gradient"
-        android:gravity="center"
-        android:padding="@dimen/dp6"
-        android:shadowColor="#000000"
-        android:shadowDx="2"
-        android:shadowDy="2"
-        android:text="Dialogue Text"
-        android:textColor="@android:color/white"
-        android:textSize="20sp" />
-
-    <LinearLayout
-        android:id="@+id/optionsLayout"
-        style="@style/ErrorButtons"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="6dp"
-        android:layout_marginTop="6dp"
-        android:layout_marginEnd="6dp"
-        android:orientation="vertical" />
-
-    <FrameLayout
-        android:id="@+id/settings_frag_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-</RelativeLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
